done backend

+++ src/controllers/menu-Controller.js
@@ -11,7 +11,7 @@ const {
 /////////////////   create  ////////////////
 exports.menuCreate = async (req, res, next) => {
   try {
-   
+  
     const menu = req.body;
 
   
@@ -44,10 +44,10 @@ exports.menuCreate = async (req, res, next) => {
 /////////////////   update  ////////////////
 exports.updateMenus = async (req, res, next) => {
   try {
+   
     const { value, error } = checkMenuIdSchema.validate(req.body);
     // const updateMenu = req.body
-    console.log(value)
-    console.log(error)
+   
     if (error) {
       return next(createError("cannot update menu", 400));
     }
@@ -72,6 +72,10 @@ exports.updateMenus = async (req, res, next) => {
     res.status(200).json({ nessage: "update menu successed", updateMenu });
   } catch (error) {
     next(error);
+  }finally{
+    if(req.file){
+      fs.unlink(req.file.path)
+    }
   }
 };
 
@@ -79,13 +83,25 @@ exports.updateMenus = async (req, res, next) => {
 exports.deleteMenu = async (req, res, next) => {
   try {
     req.params.deleteMenu = +req.params.deleteMenu;
-
+    console.log(req.params.deleteMenu)
     const { value, error } = checkMenuIdForDelete.validate(req.params);
 
-    console.log(error);
+    
     if (error) {
       return next(createError("can not delete menu", 400));
     }
+
+    const isExistData = await prisma.orderMenu.findFirst({
+      where : {
+        menuId : value.deleteMenu
+      }
+    })
+
+    if(isExistData)
+    {
+      return next(createError('This Menu already in use',400))
+    }
+
     await prisma.menus.delete({
       where: {
         id: value.deleteMenu,
@@ -109,18 +125,30 @@ exports.getMenus = async (req,res,next) => {
       role: req.user.role,
     };
 
-    const response = {
-      message: "Get all menu success",
-      data: {
-        user,
-        menus: getAllMenu,
-      },
-    };
-    res.status(201).json({ response });
+
+    res.status(201).json({ user,getAllMenu });
   } catch (error) {
     next(error);
   }
 };
+
+//////////////////  GET MENU BY ID ////////////
+exports.getMenuById = async (req,res,next) => {
+  try {
+         const data = req.params.menuId
+         console.log('data',data)
+         const getMenuId = await prisma.menus.findFirst({
+          where : {
+            id : +data
+          }
+         })
+         res.status(201).json({ getMenuId });
+
+  } catch (error) {
+    console.log(error)
+    next(error)
+  }
+}
